flask_ui/pyproject.toml
[project]
name = "flask-ui"
version = "0.1.0"
description = "Flask UI for SDN API"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "gunicorn>=23.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []